+++ .pylintrc
@@ -320,7 +320,7 @@ exclude-protected=_asdict,_fields,_replace,_source,_make
 [DESIGN]
 
 # Maximum number of arguments for function / method
-max-args=6
+max-args=7
 
 # Argument names that match this expression will be ignored. Default to name
 # with leading underscore
