+++ gpkit/tools/__init__.py
@@ -3,4 +3,3 @@ from .tools import te_exp_minus1
 from .tools import composite_objective
 from .tools import mdparse
 from .tools import mdmake
-from .tools import BoundedConstraintSet
